WIP add logs to the test
